+++ src/Auth/AuthorizesApiUser.php
@@ -14,8 +14,10 @@ declare(strict_types=1);
 
 namespace Konekt\BearerAuth\Auth;
 
+use Illuminate\Auth\Events\Login;
 use Illuminate\Contracts\Auth\Guard;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Event;
 use Konekt\BearerAuth\Exceptions\ApiAuthorizationException;
 
 trait AuthorizesApiUser
@@ -58,8 +60,32 @@ trait AuthorizesApiUser
         }
     }
 
+    private function fireLoginEvent()
+    {
+        if (config('konekt.bearer_auth.fire_login_event')) {
+            Event::dispatch(
+                new Login(
+                    $this->guardName(true),
+                    $this->bearerAuth()->user(),
+                    false
+                )
+            );
+        }
+
+        if (!$this->bearerAuth()->user()->type->isApi()) {
+            throw new ApiAuthorizationException(403, 'You are no longer an API user');
+        }
+    }
+
     private function bearerAuth(): Guard
     {
-        return Auth::guard(config('konekt.bearer_auth.guard_name'));
+        return Auth::guard($this->guardName());
+    }
+
+    private function guardName(bool $fetchDefault = false): ?string
+    {
+        $default = $fetchDefault ? config('auth.defaults.guard') : null;
+
+        return config('konekt.bearer_auth.guard_name', $default);
     }
 }